(chores) camel-test-infra-activemq: cleanups
- Remove duplicated code for the embedded activeMQ
- Remove commented code
- Cleanup the initialization of embedded broker services
