+++ test-infra/camel-test-infra-activemq/src/test/java/org/apache/camel/test/infra/activemq/services/ActiveMQEmbeddedService.java
@@ -17,171 +17,39 @@
 
 package org.apache.camel.test.infra.activemq.services;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
-
-import javax.jms.ConnectionFactory;
-import javax.management.MalformedObjectNameException;
-import javax.management.ObjectName;
-
-import org.apache.activemq.ActiveMQConnectionFactory;
 import org.apache.activemq.broker.BrokerService;
-import org.apache.activemq.broker.jmx.DestinationViewMBean;
-import org.apache.camel.test.infra.activemq.common.ConnectionFactoryHelper;
-import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.junit.jupiter.api.Assertions.fail;
-
-public class ActiveMQEmbeddedService implements ActiveMQService, ConnectionFactoryAware, BeforeEachCallback, AfterEachCallback {
+/**
+ * A regular embedded broker that uses the standard transport protocols (i.e.: AMQP, OpenWire, etc)
+ */
+public class ActiveMQEmbeddedService extends AbstractActiveMQEmbeddedService {
     private static final Logger LOG = LoggerFactory.getLogger(ActiveMQEmbeddedService.class);
-    private final BrokerService brokerService;
-    private final boolean recycle;
 
     public ActiveMQEmbeddedService() {
         this(ActiveMQEmbeddedServiceBuilder.defaultBroker().brokerService());
     }
 
     public ActiveMQEmbeddedService(BrokerService brokerService) {
-        this.brokerService = brokerService;
-        this.recycle = false;
+        this(brokerService, false);
     }
 
     public ActiveMQEmbeddedService(BrokerService brokerService, boolean recycle) {
-        this.brokerService = brokerService;
-        this.recycle = recycle;
+        super(brokerService, recycle);
+    }
+
+    public String getVmURL() {
+        return getVmURL(true, false);
     }
 
     @Override
-    public void initialize() {
-        LOG.info("Trying to start the embedded ActiveMQ");
-        try {
-            brokerService.start();
-            brokerService.waitUntilStarted();
-            LOG.info("Embedded ActiveMQ running at {}", serviceAddress());
-        } catch (Exception e) {
-            LOG.warn("Unable to start embedded ActiveMQ broker: {}", e.getMessage(), e);
-            fail(e.getMessage());
-        }
-    }
-
-    @Override
-    public void shutdown() {
-        LOG.debug("Trying to stop the embedded ActiveMQ");
-        try {
-            brokerService.stop();
-            brokerService.waitUntilStopped();
-            LOG.debug("Embedded ActiveMQ stopped");
-        } catch (Exception e) {
-            LOG.warn("Error stopping embedded ActiveMQ broker: {}", e.getMessage(), e);
-        }
-    }
-
-    public void restart() {
-        shutdown();
-
-        LOG.info("Trying to start the restart ActiveMQ");
-        try {
-            brokerService.start(true);
-            brokerService.waitUntilStarted();
-            LOG.info("Embedded ActiveMQ running at {}", serviceAddress());
-        } catch (Exception e) {
-            LOG.warn("Unable to start embedded ActiveMQ broker: {}", e.getMessage(), e);
-            fail(e.getMessage());
-        }
+    public String getVmURL(boolean create) {
+        return null;
     }
 
     @Override
     public String serviceAddress() {
-        return getBrokerUri(brokerService, 0);
-    }
-
-    @Override
-    public String userName() {
-        return null;
-    }
-
-    @Override
-    public String password() {
-        return null;
-    }
-
-    public int getConnectionCount() {
-        return brokerService.getTransportConnectors().get(0).getConnections().size();
-    }
-
-    public BrokerService getBrokerService() {
-        return brokerService;
-    }
-
-    public static String getBrokerUri(BrokerService broker, int connector) {
-        try {
-            return broker.getTransportConnectors().get(connector).getPublishableConnectString();
-        } catch (Exception e) {
-            LOG.warn("Unable to get ActiveMQ broker address: {}", e.getMessage(), e);
-            return null;
-        }
-    }
-
-    public int getPort() {
-        try {
-            return brokerService.getTransportConnectors().get(0).getServer().getSocketAddress().getPort();
-        } catch (URISyntaxException | IOException e) {
-            LOG.error("Error getting the port: {}", e.getMessage());
-            throw new RuntimeException("Error getting the port", e);
-        }
-    }
-
-    @Override
-    public void afterEach(ExtensionContext extensionContext) throws Exception {
-        if (recycle) {
-            shutdown();
-        }
-
-    }
-
-    @Override
-    public void beforeEach(ExtensionContext extensionContext) throws Exception {
-        if (recycle) {
-            initialize();
-        }
-    }
-
-    public String getVmURL() {
-        return getVmURL(true);
-    }
-
-    public String getVmURL(boolean failoverURL) {
-        return failoverURL
-                ? String.format("failover:(%s?create=false)", brokerService.getVmConnectorURI().toString())
-                : this.brokerService.getVmConnectorURI().toString() + "?create=false";
-    }
-
-    public DestinationViewMBean getQueueMBean(String queueName) throws MalformedObjectNameException {
-        return getDestinationMBean(queueName, false);
-    }
-
-    public DestinationViewMBean getDestinationMBean(String destinationName, boolean topic) throws MalformedObjectNameException {
-        String domain = "org.apache.activemq";
-        String destinationType = topic ? "Topic" : "Queue";
-        ObjectName name = new ObjectName(
-                String.format("%s:type=Broker,brokerName=localhost,destinationType=%s,destinationName=%s",
-                        domain, destinationType, destinationName));
-        return (DestinationViewMBean) brokerService.getManagementContext().newProxyInstance(name,
-                DestinationViewMBean.class, true);
-    }
-
-    @Deprecated
-    public ConnectionFactory createConnectionFactory() {
-        return createConnectionFactory(null);
-    }
-
-    @Deprecated
-    public ConnectionFactory createConnectionFactory(Integer maximumRedeliveries) {
-        ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory(getVmURL());
-        return ConnectionFactoryHelper.createConnectionFactory(connectionFactory, maximumRedeliveries);
+        return getBrokerUri(0);
     }
 }
